Revert "Revert "Complete A-Gradle""
This reverts commit 3cb1778bf2a0ad95383aa3037b1d5fce476c98d7.
