+++ src/main/java/duke/Parser.java
@@ -1,9 +1,18 @@
-package main.java.duke;
+package duke;
 
-import main.java.duke.command.*;
+import duke.command.Command;
+import duke.command.DoneCommand;
+import duke.command.DeleteCommand;
+import duke.command.ByeCommand;
+import duke.command.EventCommand;
+import duke.command.DeadlineCommand;
+import duke.command.FindCommand;
+import duke.command.InvalidCommand;
+import duke.command.ListCommand;
+import duke.command.ToDoCommand;
 
 /**
- * Parser reads the input by the user and returns the relevant command.
+ * Parser reads the input by the user and returns the relevant duke.command.
  *
  * @author Zhen Xuan (Tutorial Group W12)
  * @version CS2103T AY21/22 S1
@@ -11,10 +20,10 @@ import main.java.duke.command.*;
 public class Parser {
 
     /**
-     * Reads the fullCommand and return the relevant command.
+     * Reads the fullCommand and return the relevant duke.command.
      *
      * @param fullCommand the fullCommand
-     * @return the relevant command
+     * @return the relevant duke.command
      */
     protected static Command parse(String fullCommand) {
         String[] commandPair = fullCommand.split(" ", 2);
