+++ src/main/java/duke/Ui.java
@@ -1,4 +1,4 @@
-package main.java.duke;
+package duke;
 
 import java.io.BufferedReader;
 import java.io.IOException;
