+++ src/main/java/duke/command/Command.java
@@ -1,8 +1,8 @@
-package main.java.duke.command;
+package duke.command;
 
-import main.java.duke.storage.Storage;
-import main.java.duke.tasklist.TaskList;
-import main.java.duke.Ui;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.Ui;
 
 /**
  * This is an abstract class of all the Commands that Duke can perform.
@@ -17,25 +17,25 @@ public abstract class Command {
     /**
      * Constructor.
      *
-     * @param description it should contain more instructions for the command
+     * @param description it should contain more instructions for the duke.command
      */
     public Command(String description) {
         this.description = description;
     }
 
     /**
-     * Adds the relevant actions pertaining to the command.
+     * Adds the relevant actions pertaining to the duke.command.
      *
-     * @param tasks   the task list
+     * @param tasks   the duke.task list
      * @param ui      the ui
-     * @param storage the storage for the saved task list
+     * @param storage the storage for the saved duke.task list
      */
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
 
     /**
-     * Returns true if it is a bye command and returns false otherwise
+     * Returns true if it is a bye duke.command and returns false otherwise
      *
-     * @return boolean of whether this is the last command or not
+     * @return boolean of whether this is the last duke.command or not
      */
     public boolean isExit() {
         return false;
