+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,16 +1,16 @@
-package main.java.duke.command;
+package duke.command;
 
-import main.java.duke.DukeException;
-import main.java.duke.storage.Storage;
-import main.java.duke.tasklist.TaskList;
-import main.java.duke.Ui;
-import main.java.duke.task.Deadline;
+import duke.DukeException;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.Ui;
+import duke.task.Deadline;
 
 import java.io.IOException;
 import java.time.DateTimeException;
 
 /**
- * DeadlineCommand is a command which adds a deadline task to the task list.
+ * DeadlineCommand is a duke.command which adds a deadline duke.task to the duke.task list.
  *
  * @author Zhen Xuan (Tutorial Group W12)
  * @version CS2103T AY21/22 S1
@@ -27,11 +27,11 @@ public class DeadlineCommand extends Command {
     }
 
     /**
-     * Adds the deadline into the task list.
+     * Adds the deadline into the duke.task list.
      *
-     * @param tasks   the task list
+     * @param tasks   the duke.task list
      * @param ui      the ui
-     * @param storage the storage for the saved task list
+     * @param storage the storage for the saved duke.task list
      * @throws DukeException if the date/time format is wrong or if the saved data is deleted midway /
      *                       the data file is missing by other factors
      */
@@ -40,14 +40,14 @@ public class DeadlineCommand extends Command {
         try {
             String[] deadlinePair = description.split("/by", 2);
             if (deadlinePair.length < 2 || deadlinePair[0].equals("") || deadlinePair[1].equals("")) {
-                throw new DukeException("Your add deadline command is incomplete.");
+                throw new DukeException("Your add deadline duke.command is incomplete.");
             }
             String desc = deadlinePair[0].trim();
             String date = deadlinePair[1].trim();
             tasks.addTask(new Deadline(false, desc, date));
             storage.add("D", desc, date);
         } catch (IOException e) {
-            throw new DukeException("There is an error in adding the Deadline task to your saved data.");
+            throw new DukeException("There is an error in adding the Deadline duke.task to your saved data.");
         } catch (DateTimeException e) {
             throw new DukeException("Please provide the time in YYYY-MM-DD HHMM format instead.");
         }
