+++ src/main/java/duke/command/EventCommand.java
@@ -1,16 +1,16 @@
-package main.java.duke.command;
+package duke.command;
 
-import main.java.duke.DukeException;
-import main.java.duke.storage.Storage;
-import main.java.duke.tasklist.TaskList;
-import main.java.duke.Ui;
-import main.java.duke.task.Event;
+import duke.DukeException;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.Ui;
+import duke.task.Event;
 
 import java.io.IOException;
 import java.time.DateTimeException;
 
 /**
- * EventCommand is a command which adds an event task to the task list.
+ * EventCommand is a duke.command which adds an event duke.task to the duke.task list.
  *
  * @author Zhen Xuan (Tutorial Group W12)
  * @version CS2103T AY21/22 S1
@@ -27,11 +27,11 @@ public class EventCommand extends Command {
     }
 
     /**
-     * Adds the event into the task list.
+     * Adds the event into the duke.task list.
      *
-     * @param tasks   the task list
+     * @param tasks   the duke.task list
      * @param ui      the ui
-     * @param storage the storage for the saved task list
+     * @param storage the storage for the saved duke.task list
      * @throws DukeException if the date/time format is wrong or if the saved data is deleted midway /
      *                       the data file is missing by other factors
      */
@@ -40,14 +40,14 @@ public class EventCommand extends Command {
         try {
             String[] eventPair = description.split("/at", 2);
             if (eventPair.length < 2 || eventPair[0].equals("") || eventPair[1].equals("")) {
-                throw new DukeException("Your add event command is incomplete.");
+                throw new DukeException("Your add event duke.command is incomplete.");
             }
             String desc = eventPair[0].trim();
             String date = eventPair[1].trim();
             tasks.addTask(new Event(false, desc, date));
             storage.add("E", desc, date);
         } catch (IOException e) {
-            throw new DukeException("There is an error in adding the Event task to your saved data.");
+            throw new DukeException("There is an error in adding the Event duke.task to your saved data.");
         } catch (DateTimeException e) {
             throw new DukeException("Please provide the time in YYYY-MM-DD HHMM format instead.");
         }
