+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,15 +1,15 @@
-package main.java.duke.command;
+package duke.command;
 
-import main.java.duke.DukeException;
-import main.java.duke.storage.Storage;
-import main.java.duke.tasklist.TaskList;
-import main.java.duke.Ui;
-import main.java.duke.task.ToDo;
+import duke.DukeException;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.Ui;
+import duke.task.ToDo;
 
 import java.io.IOException;
 
 /**
- * ToDoCommand is a command which adds a ToDo task to the task list.
+ * ToDoCommand is a duke.command which adds a ToDo duke.task to the duke.task list.
  *
  * @author Zhen Xuan (Tutorial Group W12)
  * @version CS2103T AY21/22 S1
@@ -26,11 +26,11 @@ public class ToDoCommand extends Command {
     }
 
     /**
-     * Adds the ToDo into the task list.
+     * Adds the ToDo into the duke.task list.
      *
-     * @param tasks   the task list
+     * @param tasks   the duke.task list
      * @param ui      the ui
-     * @param storage the storage for the saved task list
+     * @param storage the storage for the saved duke.task list
      * @throws DukeException if the saved data is deleted midway /
      *                       the data file is missing by other factors
      */
@@ -38,13 +38,13 @@ public class ToDoCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
             if (description.replace(" ", "").equals("")) {
-                throw new DukeException("The description for the Todo task cannot be empty.");
+                throw new DukeException("The description for the Todo duke.task cannot be empty.");
             }
             String desc = description.trim();
             tasks.addTask(new ToDo(false, desc));
             storage.add("T", desc, "");
         } catch (IOException e) {
-            throw new DukeException("There is an error in adding the Todo task to your saved data.");
+            throw new DukeException("There is an error in adding the Todo duke.task to your saved data.");
         }
     }
 }
