+++ src/main/java/duke/storage/Storage.java
@@ -1,6 +1,6 @@
-package main.java.duke.storage;
+package duke.storage;
 
-import main.java.duke.task.Task;
+import duke.task.Task;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -22,7 +22,7 @@ public interface Storage {
     ArrayList<Task> load() throws IOException;
 
     /**
-     * Changea the string in the storage file to indicate that the task is done.
+     * Changea the string in the storage file to indicate that the duke.task is done.
      *
      * @param index the position of the item
      * @throws IOException if there is an error writing to the file
@@ -30,19 +30,19 @@ public interface Storage {
     void setDone(int index) throws IOException;
 
     /**
-     * Adds a new task to the storage file.
+     * Adds a new duke.task to the storage file.
      *
-     * @param type        the type of the task
-     * @param description the description of the task
-     * @param date        the date of the task (if deadline or event)
+     * @param type        the type of the duke.task
+     * @param description the description of the duke.task
+     * @param date        the date of the duke.task (if deadline or event)
      * @throws IOException if there is an error writing to the file
      */
     void add(String type, String description, String date) throws IOException;
 
     /**
-     * Deletes the task from the storage file.
+     * Deletes the duke.task from the storage file.
      *
-     * @param index the position of the task in the list
+     * @param index the position of the duke.task in the list
      * @throws IOException if there is an error writing to the file
      */
     void delete(int index) throws IOException;
