+++ src/main/java/duke/task/Task.java
@@ -1,9 +1,9 @@
-package main.java.duke.task;
+package duke.task;
 
 import java.time.LocalDate;
 
 /**
- * A task contains its description and information of whether it is done or not.
+ * A duke.task contains its description and information of whether it is done or not.
  *
  * @author Zhen Xuan (Tutorial Group W12)
  * @version CS2103T AY21/22 S1
@@ -13,7 +13,7 @@ public abstract class Task {
     private boolean isDone;
 
     /**
-     * Constructor for task.
+     * Constructor for duke.task.
      *
      * @param description the description
      */
@@ -23,7 +23,7 @@ public abstract class Task {
     }
 
     /**
-     * Returns the status icon of the task.
+     * Returns the status icon of the duke.task.
      *
      * @return status icon
      */
@@ -32,16 +32,16 @@ public abstract class Task {
     }
 
     /**
-     * Change task to done.
+     * Change duke.task to done.
      */
     public void setDone() {
         isDone = true;
     }
 
     /**
-     * Returns the string representation of the task.
+     * Returns the string representation of the duke.task.
      *
-     * @return the string representation of the task
+     * @return the string representation of the duke.task
      */
     @Override
     public String toString() {
@@ -49,7 +49,7 @@ public abstract class Task {
     }
 
     /**
-     * Returns if the task date is equal to the date provided.
+     * Returns if the duke.task date is equal to the date provided.
      *
      * @param date the date provided
      * @return true if they are both equal
@@ -57,9 +57,9 @@ public abstract class Task {
     public abstract boolean onDate(LocalDate date);
 
     /**
-     * Returns true if the task's description contains the string
+     * Returns true if the duke.task's description contains the string
      * @param str the substring to be searched
-     * @return true if the task's description contains the string
+     * @return true if the duke.task's description contains the string
      */
     public boolean containString(String str) {
         return DESCRIPTION.contains(str);
