+++ src/test/java/duke/command/ToDoCommandTest.java
@@ -1,10 +1,9 @@
-package command;
+package duke.command;
 
 
-import main.java.duke.Ui;
-import main.java.duke.command.ToDoCommand;
-import main.java.duke.storage.StorageStub;
-import main.java.duke.tasklist.TaskListStub;
+import duke.Ui;
+import duke.storage.StorageStub;
+import duke.tasklist.TaskListStub;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
