+++ src/test/java/duke/task/DeadlineTest.java
@@ -1,6 +1,5 @@
-package task;
+package duke.task;
 
-import main.java.duke.task.Deadline;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
