+++ src/test/java/duke/task/ToDoTest.java
@@ -1,6 +1,4 @@
-package task;
-
-import main.java.duke.task.ToDo;
+package duke.task;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
