+++ text-ui-test/EXPECTED.TXT
@@ -16,25 +16,25 @@
 	____________________________________________________________
 
 	____________________________________________________________
-	 Got it. I've added this task:
+	 Got it. I've added this duke.task:
 	   [T][ ] cry
-	 Now you have 1 task in the list.
+	 Now you have 1 duke.task in the list.
 	____________________________________________________________
 
 	____________________________________________________________
-	 Got it. I've added this task:
+	 Got it. I've added this duke.task:
 	   [D][ ] die (by: tdy)
 	 Now you have 2 tasks in the list.
 	____________________________________________________________
 
 	____________________________________________________________
-	 Got it. I've added this task:
+	 Got it. I've added this duke.task:
 	   [E][ ] live (at: tmr)
 	 Now you have 3 tasks in the list.
 	____________________________________________________________
 
 	____________________________________________________________
-	 Nice! I've marked this task as done:
+	 Nice! I've marked this duke.task as done:
 	   [D][X] die (by: tdy)
 	____________________________________________________________
 
@@ -46,7 +46,7 @@
 	____________________________________________________________
 
 	____________________________________________________________
-	 Noted. I've removed this task:
+	 Noted. I've removed this duke.task:
 	   [T][ ] cry
 	____________________________________________________________
